trying to refactor out form but cannot get it to work
